pi_headless/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.24.0
viam-sdk>=0.70.0
